traci: create TraCIScenarioManagerBase
This is a preparation step before moving features to separate classes.
The aim is to break no existing code!
